Conhecendo Stores e hierarquias em REST

+++ api/rotas/fornecedores/produtos/ModeloTabelaProduto.js
@@ -7,7 +7,7 @@ const colunas = {
         allowNull: false
     },
     preco: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DOUBLE,
         allowNull: false
     },
     estoque: {
